requirements.txt
fire==0.6.0
matplotlib==3.8.0
numpy==1.26.4
PyYAML==6.0.1
scipy==1.13.1
torch==2.3.0
tqdm==4.65.0
